Gates

+++ app/Http/Controllers/UserController.php
@@ -8,12 +8,15 @@ use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\UserCreateRequest;
 use App\Http\Resources\UserResource;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view', 'users');
+
         $users = User::paginate(); 
 
         return UserResource::collection($users);
@@ -21,17 +24,27 @@ class UserController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view', 'users');
+
         $user = User::findOrFail($id);
 
         return new UserResource($user);
     }
     public function user()
     {
-        return new UserResource(Auth::user());
+        $user = Auth::user();
+
+        return (new UserResource($user))->additional([
+            'data' => [
+                'permissions' => $user->permissions()
+            ]
+        ]);
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'users');
+
         $request->validate([
             'first_name' => 'required',
             'last_name' => 'required',
@@ -52,6 +65,7 @@ class UserController extends Controller
 
     public function update($id, Request $request)
     {
+        Gate::authorize('edit', 'users');
 
         $request->validate([
             'email' => 'email',
@@ -80,7 +94,6 @@ class UserController extends Controller
 
     public function updatePassword($id, Request $request)
     {
-
         $request->validate([
             'password' => 'required',
             'password_confirm' => 'required|same:password',
@@ -97,6 +110,9 @@ class UserController extends Controller
 
     public function destroy($id)
     {
+
+        Gate::authorize('edit', 'users');
+
         User::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
